requirements.txt
streamlit ==  1.11.0
pandas == 1.4.3
numpy ==  1.23.1
Pillow== 9.2.0
huggingface_hub 
opencv-python-headless==4.4.0.42
pickleshare ==  0.7.5
requests ==  2.24.0
scikit-image == 0.19.3
streamlit_option_menu